Create user layout and home page layout
