+++ src/base/routes/index.tsx
@@ -1,8 +1,7 @@
-import React from 'react';
-
 import { lazy } from 'react';
 import { RouteObject } from 'react-router-dom';
 import HomePage from '@pages/HomePage';
+import UserLayout from '@base/layout/UserLayout';
 const PageLayout = lazy(() => import('@base/PageLayout'));
 
 const SignIn = lazy(() => import('@pages/SignIn'));
@@ -11,7 +10,13 @@ const SignUp = lazy(() => import('@pages/SignUp'));
 export const routes: RouteObject[] = [
   {
     path: '/home',
-    element: <HomePage />,
+    element: <UserLayout />,
+    children: [
+      {
+        index: true,
+        element: <HomePage />,
+      },
+    ],
   },
   {
     path: '*',
