+++ src/pages/HomePage/Schedule/Board/index.tsx
@@ -1,0 +1,3 @@
+export default function Board() {
+  return <div className='bg-gray-500 flex-1 h-full'>Board</div>;
+}
