--- src/pages/HomePage/Schedule/Sidebar/Board/index.tsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export default function Board() {
-  return <div>Board</div>;
-}
