+++ src/pages/HomePage/Schedule/Sidebar/index.tsx
@@ -1,5 +1,7 @@
-import React from 'react';
-
 export default function Sidebar() {
-  return <div>Sidebar</div>;
+  return (
+    <div className='w-[280px] min-w-[68px] bg-white h-full absolute top-0 left-0 rounded-t-2xl flex flex-col overflow-hidden'>
+      <div className='h-12 bg-gray-300 px-4 pt-2'>Sidebar</div>
+    </div>
+  );
 }
